# health-tracker-docs/yuanqi/health-tracker-openapi.yaml
openapi: 3.0.0
info:
  title: 健康管理小程序 API
  version: 1.0.0
  description: 供腾讯元器插件导入的 OpenAPI 文档（简化版）
servers:
  - url: "http://139.196.155.234:8080"
    description: "生产环境 API 服务器"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
...
    AiChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        imageUrl:
          type: string
        audioUrl:
          type: string
        store:
          type: boolean
security:
  - bearerAuth: []
paths:
...
  /api/ai/chat:
    post:
      summary: AI 对话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/ai/history:
    get:
      summary: AI 历史记录
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/ai/upload:
    post:
      summary: AI 文件上传
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...